create model Messages and RespondedApplication

+++ src/main/java/com/example/jobsearch/model/Vacancies.java
@@ -7,7 +7,7 @@ public class Vacancies {
     private String name;
     private String description;
     private Integer categoryId;
-    private Integer salary;
+    private Double salary;
     private Integer expFrom;
     private Integer expTo;
     private Boolean isActivate;
